Reorganized the package structure to make place for other environmental tools.
