+++ packages/ckeditor5-dev-env/lib/release-tools/changelog/templates/footer.hbs
