+++ packages/ckeditor5-dev-env/lib/release-tools/utils/getnewreleasetype.js
