+++ packages/ckeditor5-dev-env/tests/release-tools/fixtures/basic/package.json
