+++ packages/ckeditor5-dev-env/tests/release-tools/fixtures/basic/packages/ckeditor5-core/package.json
