+++ packages/ckeditor5-dev-env/tests/release-tools/utils/releasevalidator.js
@@ -11,21 +11,21 @@ const expect = require( 'chai' ).expect;
 const sinon = require( 'sinon' );
 const { tools } = require( '@ckeditor/ckeditor5-dev-utils' );
 
-describe( 'utils', () => {
+describe( 'dev-env/release-tools/utils', () => {
 	let validator, shExecStub;
 
 	describe( 'validator', () => {
 		beforeEach( () => {
 			shExecStub = sinon.stub( tools, 'shExec' );
 
-			validator = require( '../../lib/utils/releasevalidator' );
+			validator = require( '../../../lib/release-tools/utils/releasevalidator' );
 		} );
 
 		afterEach( () => {
 			shExecStub.restore();
 		} );
 
-		describe( 'checkOptions', () => {
+		describe( 'checkOptions()', () => {
 			it( 'throws an error when given options are invalid', () => {
 				expect( () => {
 					validator.checkOptions( {} );
@@ -39,7 +39,7 @@ describe( 'utils', () => {
 			} );
 		} );
 
-		describe( 'checkBranch', () => {
+		describe( 'checkBranch()', () => {
 			it( 'does not throw if on master and master is clean', () => {
 				shExecStub.returns( '## master...origin/master' );
